changing names to experiment_settings.py following design doc. move set_values and initailizeCombo in presenter
